+++ app/src/main/java/org/andreip/utils/Interval.java
@@ -1,26 +1,26 @@
 package org.andreip.utils;
 
-public record Interval(double min, double max) {
+public record Interval(float min, float max) {
     public static final Interval EMPTY = new Interval();
-    public static final Interval UNIVERSE = new Interval(Double.MIN_VALUE, Double.MAX_VALUE);
+    public static final Interval UNIVERSE = new Interval(Float.MIN_VALUE, Float.MAX_VALUE);
 
     public Interval() {
-        this(Double.MAX_VALUE, Double.MIN_VALUE);
+        this(Float.MAX_VALUE, Float.MIN_VALUE);
     }
 
-    public double size() {
+    public float size() {
         return max - min;
     }
 
-    public boolean contains(double x) {
+    public boolean contains(float x) {
         return min <= x && x <= max;
     }
 
-    public boolean surrounds(double x) {
+    public boolean surrounds(float x) {
         return min < x && x < max;
     }
 
-    public double clamp(double x) {
+    public float clamp(float x) {
         if (x < min) return min;
         if (x > max) return max;
         return x;
